+++ src/main/java/com/guxingke/intellij/plugin/Const.java
@@ -1,0 +1,7 @@
+package com.guxingke.intellij.plugin;
+
+public interface Const {
+
+  String CLS_JAVA_UTIL_COLLECTION = "java.util.Collection";
+  String CLS_JAVA_UTIL_STREAM_STREAM = "java.util.stream.Stream";
+}