+++ src/main/java/com/guxingke/intellij/plugin/action/HelloAction.java
@@ -1,0 +1,12 @@
+package com.guxingke.intellij.plugin.action;
+
+import com.intellij.openapi.actionSystem.AnAction;
+import com.intellij.openapi.actionSystem.AnActionEvent;
+
+public class HelloAction extends AnAction {
+
+  @Override
+  public void actionPerformed(AnActionEvent e) {
+    // TODO: insert action logic here
+  }
+}